+++ src/Stash/Interfaces/PoolInterface.php
@@ -66,7 +66,7 @@ interface PoolInterface
      * @param  array     $keys
      * @return \Iterator
      */
-    public function getItemIterator($keys);
+    public function getItems($keys);
 
     /**
      * Empties the entire cache pool of all Items.
@@ -127,4 +127,43 @@ interface PoolInterface
      * @return bool
      */
     public function setLogger($logger);
+
+    /**
+    * Forces any save-deferred objects to get flushed to the backend drivers.
+    */
+    public function commit();
+
+    /**
+    * Sets an Item to be saved at some point. This allows buffering for
+    * multi-save events.
+    *
+    * @param CacheItemInterface $item
+    * @return static The invoked object.
+    */
+    public function saveDeferred($item);
+
+    /**
+    * Sets an Item to be saved immediately.
+    *
+    * @param CacheItemInterface $item
+    * @return static The invoked object.
+    */
+    public function save($item);
+
+    /**
+     * Removes multiple items from the pool.
+     *
+     * @param array $keys
+     * @return static The invoked object.
+     */
+    public function deleteItems(array $keys);
+
+
+    /**
+    * Checks for the existance of an item in the cache.
+    *
+    * @param  array|string $key    
+    * @return boolean True if item exists in the cache, false otherwise.
+    */
+    public function exists($key);
 }